Precompile kaliber/pack
