+++ package.json
@@ -19,5 +19,8 @@
   "private": false,
   "license": "MIT",
   "description": "Returns a safe stringified version of a given string",
-  "repository": "https://github.com/kaliberjs/safe-json-stringify.git"
+  "repository": "https://github.com/kaliberjs/safe-json-stringify.git",
+  "devDependencies": {
+    "@kaliber/pack": "^0.0.2"
+  }
 }
